+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `GraphicConfig.scissor_test` parameter
 - Support for Android < 9.0
 - `ngl-viewer` and `ngl-diff` support and release on Apple platforms
+- Support of OpenGL < 4.4 and OpenGLES < 3.1
 
 ## [2024.0] [libnopegl 0.11.0] - 2024-02-02
 ### Added